+++ ninedot-infrastructure/src/main/java/org/sopt36/ninedotserver/mandalart/persistence/CoreGoalRepository.java
@@ -1,14 +1,16 @@
-package org.sopt36.ninedotserver.mandalart.repository;
+package org.sopt36.ninedotserver.mandalart.persistence;
 
 import java.util.List;
 import java.util.Optional;
-import org.sopt36.ninedotserver.mandalart.domain.CoreGoal;
+import org.sopt36.ninedotserver.mandalart.model.CoreGoal;
+import org.sopt36.ninedotserver.mandalart.persistence.querydsl.CoreGoalRepositoryCustom;
+import org.sopt36.ninedotserver.mandalart.port.out.CoreGoalRepositoryPort;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface CoreGoalRepository
-    extends JpaRepository<CoreGoal, Long>, CoreGoalRepositoryCustom {
+    extends JpaRepository<CoreGoal, Long>, CoreGoalRepositoryPort, CoreGoalRepositoryCustom {
 
     int countCoreGoalByMandalartId(Long mandalartId);
