+++ ninedot-infrastructure/src/main/java/org/sopt36/ninedotserver/mandalart/persistence/HistoryRepository.java
@@ -1,14 +1,17 @@
-package org.sopt36.ninedotserver.mandalart.repository;
+package org.sopt36.ninedotserver.mandalart.persistence;
 
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
-import org.sopt36.ninedotserver.mandalart.domain.History;
+import org.sopt36.ninedotserver.mandalart.model.History;
+import org.sopt36.ninedotserver.mandalart.persistence.querydsl.HistoryRepositoryCustom;
+import org.sopt36.ninedotserver.mandalart.port.out.HistoryRepositoryPort;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HistoryRepository extends JpaRepository<History, Long>, HistoryRepositoryCustom {
+public interface HistoryRepository
+    extends JpaRepository<History, Long>, HistoryRepositoryPort, HistoryRepositoryCustom {
 
     boolean existsBySubGoalSnapshotIdAndCompletedDate(Long subGoalSnapshotId, LocalDate now);
