+++ ninedot-infrastructure/src/main/java/org/sopt36/ninedotserver/mandalart/persistence/SubGoalRepository.java
@@ -1,13 +1,14 @@
-package org.sopt36.ninedotserver.mandalart.repository;
+package org.sopt36.ninedotserver.mandalart.persistence;
 
 import java.util.Optional;
-import org.sopt36.ninedotserver.mandalart.domain.CoreGoal;
-import org.sopt36.ninedotserver.mandalart.domain.SubGoal;
+import org.sopt36.ninedotserver.mandalart.model.CoreGoal;
+import org.sopt36.ninedotserver.mandalart.model.SubGoal;
+import org.sopt36.ninedotserver.mandalart.port.out.SubGoalRepositoryPort;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SubGoalRepository extends JpaRepository<SubGoal, Long> {
+public interface SubGoalRepository extends JpaRepository<SubGoal, Long>, SubGoalRepositoryPort {
 
     int countByCoreGoalId(Long coreGoalId);
