+++ ninedot-infrastructure/src/main/java/org/sopt36/ninedotserver/mandalart/persistence/querydsl/CoreGoalRepositoryCustom.java
@@ -1,4 +1,4 @@
-package org.sopt36.ninedotserver.mandalart.repository;
+package org.sopt36.ninedotserver.mandalart.persistence.querydsl;
 
 public interface CoreGoalRepositoryCustom {
