+++ ninedot-infrastructure/src/main/java/org/sopt36/ninedotserver/mandalart/persistence/querydsl/CoreGoalSnapshotRepositoryImpl.java
@@ -1,11 +1,11 @@
-package org.sopt36.ninedotserver.mandalart.repository;
+package org.sopt36.ninedotserver.mandalart.persistence.querydsl;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.sopt36.ninedotserver.mandalart.domain.CoreGoalSnapshot;
-import org.sopt36.ninedotserver.mandalart.domain.QCoreGoal;
-import org.sopt36.ninedotserver.mandalart.domain.QCoreGoalSnapshot;
+import org.sopt36.ninedotserver.mandalart.model.CoreGoalSnapshot;
+import org.sopt36.ninedotserver.mandalart.model.QCoreGoal;
+import org.sopt36.ninedotserver.mandalart.model.QCoreGoalSnapshot;
 
 @RequiredArgsConstructor
 public class CoreGoalSnapshotRepositoryImpl implements CoreGoalSnapshotRepositoryCustom {
